rename Search.js to Searching.js
